requirements.txt
blinker==1.8.2
click==8.1.7
docx2pdf==0.1.8
fire==0.6.0
Flask==3.0.3
fonttools==4.53.1
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
numpy==2.1.0
opencv-python-headless==4.10.0.84
pdf2docx==0.5.8
pdfkit==1.0.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pypdf==4.3.1
python-docx==1.1.2
six==1.16.0
termcolor==2.4.0
tqdm==4.66.5
typing_extensions==4.12.2
Werkzeug==3.0.4
